+++ src/constants/categories.ts
@@ -1,0 +1,12 @@
+export const CATEGORY = {
+  FE: 'Frontend',
+  BE: 'Backend',
+  DO: 'DevOps',
+  QA: 'QA'
+};
+export const CATEGORY_LIST = [
+  CATEGORY.FE,
+  CATEGORY.BE,
+  CATEGORY.DO,
+  CATEGORY.QA
+];
